+++ Tests/TuistGeneratorTests/Generator/LinkGeneratorTests.swift
@@ -27,13 +27,13 @@ final class LinkGeneratorErrorTests: XCTestCase {
     func test_generateEmbedPhase() throws {
         var dependencies: [DependencyReference] = []
         dependencies.append(DependencyReference.absolute(AbsolutePath("/test.framework")))
-        dependencies.append(DependencyReference.product("waka.framework"))
+        dependencies.append(DependencyReference.product(target: "Test"))
         let pbxproj = PBXProj()
         let pbxTarget = PBXNativeTarget(name: "Test")
         let fileElements = ProjectFileElements()
         let wakaFile = PBXFileReference()
         pbxproj.add(object: wakaFile)
-        fileElements.products["waka.framework"] = wakaFile
+        fileElements.products["Test"] = wakaFile
         let sourceRootPath = AbsolutePath("/")
 
         try subject.generateEmbedPhase(dependencies: dependencies,
@@ -58,7 +58,7 @@ final class LinkGeneratorErrorTests: XCTestCase {
 
     func test_generateEmbedPhase_throws_when_aProductIsMissing() throws {
         var dependencies: [DependencyReference] = []
-        dependencies.append(DependencyReference.product("waka.framework"))
+        dependencies.append(DependencyReference.product(target: "Test"))
         let pbxproj = PBXProj()
         let pbxTarget = PBXNativeTarget(name: "Test")
         let fileElements = ProjectFileElements()
@@ -69,7 +69,7 @@ final class LinkGeneratorErrorTests: XCTestCase {
                                                             pbxproj: pbxproj,
                                                             fileElements: fileElements,
                                                             sourceRootPath: sourceRootPath)) {
-            XCTAssertEqual($0 as? LinkGeneratorError, LinkGeneratorError.missingProduct(name: "waka.framework"))
+            XCTAssertEqual($0 as? LinkGeneratorError, LinkGeneratorError.missingProduct(name: "Test"))
         }
     }
 
@@ -217,7 +217,7 @@ final class LinkGeneratorErrorTests: XCTestCase {
     func test_generateLinkingPhase() throws {
         var dependencies: [DependencyReference] = []
         dependencies.append(DependencyReference.absolute(AbsolutePath("/test.framework")))
-        dependencies.append(DependencyReference.product("waka.framework"))
+        dependencies.append(DependencyReference.product(target: "Test"))
         let pbxproj = PBXProj()
         let pbxTarget = PBXNativeTarget(name: "Test")
         let fileElements = ProjectFileElements()
@@ -225,7 +225,7 @@ final class LinkGeneratorErrorTests: XCTestCase {
         pbxproj.add(object: testFile)
         let wakaFile = PBXFileReference()
         pbxproj.add(object: wakaFile)
-        fileElements.products["waka.framework"] = wakaFile
+        fileElements.products["Test"] = wakaFile
         fileElements.elements[AbsolutePath("/test.framework")] = testFile
 
         try subject.generateLinkingPhase(dependencies: dependencies,
@@ -259,7 +259,7 @@ final class LinkGeneratorErrorTests: XCTestCase {
 
     func test_generateLinkingPhase_throws_whenProductIsMissing() throws {
         var dependencies: [DependencyReference] = []
-        dependencies.append(DependencyReference.product("waka.framework"))
+        dependencies.append(DependencyReference.product(target: "Test"))
         let pbxproj = PBXProj()
         let pbxTarget = PBXNativeTarget(name: "Test")
         let fileElements = ProjectFileElements()
@@ -268,7 +268,7 @@ final class LinkGeneratorErrorTests: XCTestCase {
                                                               pbxTarget: pbxTarget,
                                                               pbxproj: pbxproj,
                                                               fileElements: fileElements)) {
-            XCTAssertEqual($0 as? LinkGeneratorError, LinkGeneratorError.missingProduct(name: "waka.framework"))
+            XCTAssertEqual($0 as? LinkGeneratorError, LinkGeneratorError.missingProduct(name: "Test"))
         }
     }
 
@@ -435,7 +435,7 @@ final class LinkGeneratorErrorTests: XCTestCase {
     func createProjectFileElements(for targets: [Target]) -> ProjectFileElements {
         let projectFileElements = ProjectFileElements(playgrounds: MockPlaygrounds())
         targets.forEach {
-            projectFileElements.products[$0.productNameWithExtension] = PBXFileReference(path: $0.productNameWithExtension)
+            projectFileElements.products[$0.name] = PBXFileReference(path: $0.productNameWithExtension)
         }
 
         return projectFileElements